fix: complaints saving, alerts when logging out and more

+++ src/activities/agent/home.js
@@ -5,7 +5,8 @@ import {
   StyleSheet,
   View,
   AsyncStorage,
-  Dimensions
+  Dimensions,
+  Alert
 } from "react-native";
 
 import { Appbar, ProgressBar, Colors } from "react-native-paper";
@@ -55,8 +56,25 @@ class Screen extends React.Component {
           <Appbar.Action
             icon="power-settings-new"
             onPress={() => {
-              AsyncStorage.clear();
-              this.props.navigation.navigate("DriverLogin");
+              Alert.alert(
+                "Log Out?",
+                `Are you sure you remove all the data on the app?`,
+                [
+                  {
+                    text: "Cancel",
+                    onPress: () => console.log("Cancel Pressed"),
+                    style: "cancel"
+                  },
+                  {
+                    text: "OK",
+                    onPress: async () => {
+                      AsyncStorage.clear();
+                      this.props.navigation.navigate("Loading");
+                    }
+                  }
+                ],
+                { cancelable: false }
+              );
             }}
           />
         </Appbar.Header>
